Move AI modules into services
