+++ app/services/ai/vector-store.js
@@ -1,6 +1,6 @@
 const { PGVectorStore } = require('@langchain/community/vectorstores/pgvector')
 const { embeddings } = require('./ai')
-const { getConfig } = require('../config/db')
+const { getConfig } = require('../../config/db')
 
 let vectorStore
